# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cryptex Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/balance:
    get:
      operationId: getUserBalance
      tags: [user]
      summary: Get user balance
      responses:
        "200":
          description: User balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBalance"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user/profile:
    patch:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile (email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [user]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Settings (public) ─────────────────────────────────────────────────────
...
  /savings/plans:
    get:
      operationId: listSavingsPlans
      tags: [savings]
      summary: List user savings plans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, early_withdrawn]
      responses:
        "200":
          description: Savings plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsPlan"
    post:
      operationId: createSavingsPlan
      tags: [savings]
      summary: Create a savings plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsPlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsPlan"
        "400":
          description: Validation or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /savings/plans/{id}/withdraw-early:
    post:
      operationId: earlyWithdrawSavings
      tags: [savings]
      summary: Early withdraw a savings plan (forfeits profit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Early withdrawal processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsPlan"
        "400":
          description: Cannot withdraw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Withdraw ───────────────────────────────────────────────────────────────
  /withdraw/requests:
    get:
      operationId: listWithdrawals
      tags: [withdraw]
      summary: List user withdrawal requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, rejected]
      responses:
        "200":
          description: Withdrawal requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdraw]
      summary: Submit withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Transaction History ────────────────────────────────────────────────────
  /user/history:
    get:
      operationId: getTransactionHistory
      tags: [user]
      summary: Combined transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [exchange, savings, withdrawal, all]
            default: all
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistory"

  # ── Admin ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth Schemas ─────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        isAdmin:
          type: boolean
        availableEur:
          type: number
        lockedEur:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, email, isAdmin, availableEur, lockedEur, createdAt]
    ProfileUpdate:
      type: object
      properties:
        email:
          type: string
          format: email

    # ── Balance / Dashboard ───────────────────────────────────────────────────
    UserBalance:
      type: object
      properties:
        availableEur:
          type: number
        lockedEur:
          type: number
        totalEur:
          type: number
      required: [availableEur, lockedEur, totalEur]
...
    ExchangeOrder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        usdtAmount:
          type: number
        eurAmount:
          type: number
        txid:
          type: string
        status:
          type: string
          enum: [pending, completed, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, userId, usdtAmount, eurAmount, txid, status, createdAt]
...
    SavingsPlanInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 100
        durationDays:
          type: integer
          enum: [7, 14, 30]
      required: [amount, durationDays]
    SavingsPlan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        durationDays:
          type: integer
        dailyRate:
          type: number
        profitEarned:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        lastPayoutDate:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [active, completed, early_withdrawn]
      required: [id, userId, amount, durationDays, dailyRate, profitEarned, startDate, endDate, status]

    # ── Withdrawal Schemas ────────────────────────────────────────────────────
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 10
        method:
          type: string
          enum: [interac, sepa, wise, paypal, bank_wire]
        accountDetails:
          type: object
          additionalProperties: true
      required: [amount, method, accountDetails]
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
        accountDetails:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, completed, rejected]
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, userId, amount, method, accountDetails, status, createdAt]

    # ── Transaction History ────────────────────────────────────────────────────
    TransactionHistory:
      type: object
      properties:
        exchanges:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeOrder"
        savings:
          type: array
          items:
            $ref: "#/components/schemas/SavingsPlan"
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
      required: [exchanges, savings, withdrawals, total]

    # ── Admin Schemas ─────────────────────────────────────────────────────────
...